Added the map and the tiles used
Added the GBMB files for the maps and the GMTD files for the tiles. Both have the exported .z80 file. These are only here to show which format to output maps and tiles as. Unlike the GBMB and GMTD files, they do not always contain all of the necessary data.